got the update function working, including both the clientside javascript and the backend

+++ public/js/dashboard.js
@@ -1,24 +1,42 @@
-const delBtnHandler = async (event) => {
-    console.log('here')
+const multiBtnHandler = async (event) => {
     if (event.target.hasAttribute('data-id')) {
-        const id = event.target.getAttribute('data-id');
-        console.log(id);
-        const response = await fetch(`/api/posts/${id}`, {
-            method: 'DELETE',
-        });
+        if(event.target.getAttribute('id') === 'delete'){
+            const id = event.target.getAttribute('data-id');
+            const response = await fetch(`/api/posts/${id}`, {
+                method: 'DELETE',
+            });
 
-        if (response.ok) {
-            document.location.replace('/dashboard');
-        } else {
-            alert('Failed to delete blogpost')
+            if (response.ok) {
+                document.location.replace('/dashboard');
+            } else {
+                alert('Failed to delete blogpost')
+            }
+        } 
+        else if(event.target.getAttribute('id') === 'update'){
+            const id = event.target.getAttribute('data-id');
+            document.getElementById(`post${id}`).setAttribute('class', 'hidden')
+            document.getElementById(`form${id}`).setAttribute('class', 'update-form')
+
+        }
+        else if(event.target.getAttribute('id') === 'save-update'){
+            const id = event.target.getAttribute('data-id');
+            const post_text = document.querySelector(`#new-comment-${id}`).value.trim();
+            const response = await fetch(`/api/posts/${id}`, {
+                method: 'PUT',
+                body: JSON.stringify({post_text}),
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+            });
+            if (response.ok) {
+                document.location.replace('/dashboard');
+            } else {
+                alert('Failed to update blogpost')
+            }
         }
     }   
 };
 
-// const updateBtnHandler = async (event) => {
-//     event.preventDefault();
-//     const id = document.querySelector('#update').getAttribute('data-id');
-// }
 
-document.querySelector('.my-posts').addEventListener('click', delBtnHandler);
-// document.querySelector('.my-posts').addEventListener('click', updateBtnHandler);
+
+document.querySelector('.my-posts').addEventListener('click', multiBtnHandler);
